added plugin to highlight code lines
It is not typical syntax highlighting. I created this plugin to highlight lines in code blocks that were added. The idea was to avoid usage of Prism, Shiki, etc. I wanted something simple, that will not add useless code. So, the effect is that new lines have primary color, and other lines have transparency.